+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Types of changes:
 ### Added
 
 * Function/script to generate comparisons between different methods/approaches using Jaccard similarity
+* Added support for generating text files from jsonlines using `mml-jsonl-to-txt`
 
 
 ## [1.0.1] - 2024-10-16